Fixed formatting of footnotes.
The Word and ODT templates use hanging paragraphs for footnotes. We
achieve the same here by loading the footmisc package with the `hang'
option. The `multiple' option of the same package ensures commas are
placed between the footnote markers when several footnotes are given
at a single point in the text. The package also allows us to set the
footnotes half-ragged instead of fully justified.